Make organisations facet filterable
Finder facets usually need an  property, with `allowed_values`
a hard coded list of the labels/values to populate the filter
with.

The behaviour in finders is changing, so that a filter facet
without `allowed_values`  will have it's labels/values
calculated as part of the call to rummager, and will be
dynamic based on the results of the rummager call.